package renamed from logger to logging
